% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-network.R
\name{build_by_network}
\alias{build_by_network}
\title{Build a network where entities share values from another field}
\usage{
build_by_network(data, field, by, count, measure, threshold, min_occur)
}
\description{
Build a network where entities share values from another field
}
\keyword{internal}
