% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal-network.R
\name{build_windows}
\alias{build_windows}
\title{Build time window definitions}
\usage{
build_windows(min_time, max_time, window, step, strategy)
}
\description{
Build time window definitions
}
\keyword{internal}
