% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{detect_format}
\alias{detect_format}
\title{Detect bibliometric file format}
\usage{
detect_format(file)
}
\arguments{
\item{file}{Path to file.}
}
\value{
Character: format name or \code{"unknown"}.
}
\description{
Detect bibliometric file format
}
\keyword{internal}
