% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co-network.R
\name{ensure_list_column}
\alias{ensure_list_column}
\title{Ensure a column is a list-column, splitting if needed}
\usage{
ensure_list_column(data, field, sep = ";")
}
\description{
Ensure a column is a list-column, splitting if needed
}
\keyword{internal}
