% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{position_independent_counts}
\alias{position_independent_counts}
\title{List of all available counting methods}
\usage{
position_independent_counts()
}
\value{
Character vector of method names.
}
\description{
List of all available counting methods
}
\keyword{internal}
