% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bibnets_network}
\alias{print.bibnets_network}
\title{Print a bibnets network edge list}
\usage{
\method{print}{bibnets_network}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{bibnets_network} data frame.}

\item{n}{Integer. Number of rows to show. Default 10.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a bibnets network edge list
}
\examples{
data(biblio_data)
edges <- author_network(biblio_data, "collaboration")
print(edges)
}
