% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.citenets_network}
\alias{print.citenets_network}
\title{Print a citenets network edge list}
\usage{
\method{print}{citenets_network}(x, n = 10L, ...)
}
\arguments{
\item{x}{A \code{citenets_network} data frame.}

\item{n}{Integer. Number of rows to show. Default 10.}

\item{...}{Ignored.}
}
\description{
Print a citenets network edge list
}
