% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{read_single_biblio}
\alias{read_single_biblio}
\title{Read a single bibliometric file}
\usage{
read_single_biblio(file, format, id, actors, sep, ...)
}
\description{
Read a single bibliometric file
}
\keyword{internal}
