% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-biblio.R
\name{resolve_paths}
\alias{resolve_paths}
\title{Resolve file paths from a file, vector of files, or directory}
\usage{
resolve_paths(path)
}
\description{
Resolve file paths from a file, vector of files, or directory
}
\keyword{internal}
