% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_field}
\alias{split_field}
\title{Parse semicolon-delimited strings into list-column}
\usage{
split_field(x, sep = ";")
}
\arguments{
\item{x}{A character vector of semicolon-delimited strings.}

\item{sep}{Character. Delimiter. Default \code{";"}.}
}
\value{
A list of character vectors.
}
\description{
Splits semicolon-separated strings (common in Scopus/WoS exports) into
character vectors, trimming whitespace.
}
\examples{
split_field(c("Alice; Bob; Carol", "Dave; Eve"))
}
