% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bibnets_network}
\alias{summary.bibnets_network}
\title{Summarise a bibnets network}
\usage{
\method{summary}{bibnets_network}(object, ...)
}
\arguments{
\item{object}{A \code{bibnets_network} data frame.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarise a bibnets network
}
\examples{
data(biblio_data)
edges <- author_network(biblio_data, "collaboration")
summary(edges)
}
