% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.citenets_network}
\alias{summary.citenets_network}
\title{Summarise a citenets network}
\usage{
\method{summary}{citenets_network}(object, ...)
}
\arguments{
\item{object}{A \code{citenets_network} data frame.}

\item{...}{Ignored.}
}
\description{
Summarise a citenets network
}
