% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{to_tbl_graph}
\alias{to_tbl_graph}
\title{Convert edge data frame to tbl_graph}
\usage{
to_tbl_graph(edges, directed = FALSE)
}
\arguments{
\item{edges}{A data frame with at least \code{from}, \code{to}, \code{weight} columns.}

\item{directed}{Logical. Default \code{FALSE}.}
}
\value{
A tbl_graph object (tidygraph).
}
\description{
Convert edge data frame to tbl_graph
}
\examples{
\dontrun{
data(biblio_data)
edges <- keyword_network(biblio_data)
tg <- to_tbl_graph(edges)
}
}
